+++ src/commands/hardis/work/publish.ts
@@ -5,8 +5,8 @@ import { AnyJson } from '@salesforce/ts-types';
 import * as c from 'chalk';
 // import * as open from 'open';
 import * as prompts from 'prompts';
-import { execCommand, getCurrentGitBranch, git, interactiveGitAdd, uxLog } from '../../../../common/utils';
-import { getConfig } from '../../../../config';
+import { execCommand, getCurrentGitBranch, git, interactiveGitAdd, uxLog } from '../../../common/utils';
+import { getConfig } from '../../../config';
 
 // Initialize Messages with the current plugin directory
 Messages.importMessagesDirectory(__dirname);